Cleaned up code, finalize how I will do generate listener threads, added timers into code. Lily keep me posted how it works. It worked well ony my dummy program

+++ generate/socket.c
@@ -41,7 +41,7 @@ int createServerSocket(unsigned short port)
     return servSock;
 }
 
-int createClientSocket(char * ip, unsigned short port)
+int createClientSocket(char * ip, unsigned int port)
 {
     int sock;
     if ((sock = socket(AF_INET, SOCK_STREAM, 0)) < 0)
